confirm dialog on delete and snackbar (toaster) examples after delete, done, new todo creation

+++ client/src/app/todo/todo.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { TodoApi, Todo } from "app/shared/sdk";
 import { Observable } from "rxjs/Observable";
-import { TdLoadingService } from '@covalent/core';
+import { TdLoadingService, TdDialogService } from '@covalent/core';
+import { MdSnackBar } from '@angular/material';
+
 
 @Component({
   selector: 'poc-todo',
@@ -11,7 +13,12 @@ import { TdLoadingService } from '@covalent/core';
 export class TodoComponent implements OnInit {
   todos: Todo[];
 
-  constructor(private todoApi: TodoApi, private loadingService: TdLoadingService) { }
+  constructor(
+    private todoApi: TodoApi,
+    private loadingService: TdLoadingService,
+    private snackBarService: MdSnackBar,
+    private dialogService: TdDialogService
+  ) { }
 
   ngOnInit() {
     this.refreshTodoList();
@@ -22,20 +29,35 @@ export class TodoComponent implements OnInit {
     .subscribe((result) => {
       // this.refreshTodoList();
       this.todos.push(result);
+      this.snackBarService.open('Todo created', 'Ok', { duration: 2000 });
     })
   }
 
   onDeleteTodo(todo: Todo) {
-    this.todoApi.deleteById(todo.id)
-    .subscribe((result) => {
-      this.refreshTodoList();
-    })
+    this.dialogService
+      .openConfirm({message: 'Are you sure you want to delete this todo?'})
+      .afterClosed().subscribe((confirm: boolean) => {
+        if (confirm) {
+          this.loadingService.register('users.list');
+          this.todoApi.deleteById(todo.id)
+          .subscribe((result) => {
+            this.loadingService.resolve('users.list');
+            this.snackBarService.open('Todo deleted', 'Ok');
+            this.refreshTodoList();
+          });
+        }
+      });
   }
 
   onToggleDoneTodo(todo: Todo) {
     this.todoApi.patchAttributes(todo.id, {done: !todo.done})
     .subscribe((result) => {
       this.refreshTodoList();
+      this.snackBarService.open(`Todo ${result.done ? '' : 'not'} done`, 'Ok',
+      {
+        duration: 2000,
+        extraClasses: ['success']
+      });
     })
   }
